separate SideNavAnchor & TopNavAnchor into own components & correct styling on font colours
