+++ src/components/Header/TopNavAnchor.js
@@ -1,0 +1,33 @@
+import styled from 'styled-components'
+import remcalc from 'remcalc'
+import is from 'styled-is'
+
+import Anchor from '../Common/Anchor'
+import headerAnchorStyles from './headerAnchorStyles'
+
+const TopNavAnchor = styled(Anchor)`
+  ${headerAnchorStyles}
+  font-weight: 400;
+  font-size: ${remcalc(17)};
+  line-height: ${remcalc(24)};
+  opacity: 1;
+  padding: ${remcalc(10)} ${remcalc(15)} ${remcalc(14)};
+
+  color: ${props => props.theme.colors.textLight};
+
+  &.active,
+  &:hover {
+    color: ${props => props.theme.colors.text};
+  }
+
+  ${is('dark')`
+    color: ${props => props.theme.colors.white};
+    
+    &.active,
+    &:hover {
+      color: ${props => props.theme.colors.white};
+    }
+  `}
+`
+
+export default TopNavAnchor
